+++ src/libkefir_parse_ethtool.c
@@ -675,7 +675,7 @@ ethtool_compose_rule(enum ethtool_flow_type flow_type,
 }
 
 struct kefir_rule *
-kefir_parse_rule_ethtool(const char **user_rule, unsigned int rule_size)
+kefir_parse_rule_ethtool(const char **user_rule, size_t rule_size)
 {
 	struct ethtool_option current_opt = { .name = "" };
 	struct kefir_value match_val = {0};
